Rename utility method and update files
