+++ client/components/relatedProducts/Card.jsx
@@ -5,7 +5,7 @@ import { AiOutlineStar } from 'react-icons/ai';
 import StarRating from '../StarRating';
 
 import getStarRatingAvg from '../../utils/getStarRatingAvg';
-import getPhotos from '../../utils/getImageUrl';
+import getPhotos from '../../utils/getPhotos';
 
 import '../../css/relatedProducts/Card.scss';
