+++ client/tests/getPhotos.test.js
@@ -1,0 +1,7 @@
+import getPhotos from '../utils/getPhotos';
+
+describe('getPhotos utitlity function', () => {
+  test('it should return a function', () => {
+    expect(typeof getPhotos()).toBe('function');
+  });
+});
